% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASRjson.R
\docType{package}
\name{ASRjson-package}
\alias{ASRjson}
\alias{ASRjson-package}
\title{ASRjson: AlphaSimR simulations from json}
\description{
Functions for automating simulation runs designed in a GUI.
The GUI outputs a json file that the functions in this library
will use to run an AlphaSimR simulation.
}
\author{
\strong{Maintainer}: Chris Gaynor \email{gaynor.robert@hotmail.com} (\href{https://orcid.org/0000-0003-0558-6656}{ORCID})

Authors:
\itemize{
  \item Marlee Labroo
}

}
\keyword{internal}
