% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASRjson.R
\name{ASRjson}
\alias{ASRjson}
\title{Run AlphaSimR simulations from json}
\usage{
ASRjson(json, output)
}
\arguments{
\item{json}{path to input json file}

\item{output}{name of output file}
}
\description{
Reads in a file in json format and uses
the information to run an AlphaSimR simulation. Output from
the simulation is saved as an R data file.
}
