% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASRjson.R
\name{json2script}
\alias{json2script}
\title{Convert json AlphaSimR script}
\usage{
json2script(json, output)
}
\arguments{
\item{json}{path to input json file}

\item{output}{name of output file}
}
\description{
Reads in a file in json format and uses
the information create an AlphaSimR script.
}
\examples{
\dontrun{
json = system.file("extdata", "example.json",
                   package = "ASRjson")
json2script(json, "example.R")
}


}
